chore: split backend config, exclude it fro coverage
