+++ backend/src/main/kotlin/fr/gouv/dgampa/rapportnav/infrastructure/api/filter/CustomAuthenticationFilter.kt
@@ -1,4 +1,4 @@
-package fr.gouv.dgampa.rapportnav.config
+package fr.gouv.dgampa.rapportnav.infrastructure.api.filter
 
 import fr.gouv.dgampa.rapportnav.domain.use_cases.auth.TokenService
 import jakarta.servlet.FilterChain
