+++ backend/src/test/kotlin/fr/gouv/gmampa/rapportnav/config/AuditorAwareImplTest.kt
@@ -1,6 +1,6 @@
 package fr.gouv.gmampa.rapportnav.config
 
-import fr.gouv.dgampa.rapportnav.config.AuditorAwareImpl
+import fr.gouv.dgampa.rapportnav.infrastructure.database.AuditorAwareImpl
 import fr.gouv.dgampa.rapportnav.domain.entities.user.RoleTypeEnum
 import fr.gouv.dgampa.rapportnav.domain.entities.user.User
 import org.junit.jupiter.api.AfterEach
