+++ sonar-project.properties
@@ -22,3 +22,7 @@ sonar.exclusions=\
   backend/target/**,\
   backend/gradle/**,\
   backend/.gradle/**
+
+# Exclude config folder from test coverage (pure Spring wiring, no testable logic)
+sonar.coverage.exclusions=\
+  **/rapportnav/config/**
